Clean old code
- InputsOutputsSpecs:
    - remove all methods from previous api,
    - remove perturbation_domain attribute
    - move into decomon.layers.inputs_outputs_specs
- ForwardMode: removed, replaced by ibp + affine booleans
- PerturbationDomain and related stuff -> decomon.perturbation_domain
- decomon.core -> decomon.constants: contains only enumerations
- decomon.keras_utils: remove unused operations like BatchDiag and
  BatchIdentity
- decomon.utils: keep only activation relaxations, moved into
  decomon.layers.activations.utils. Remove get_linear_hull_s_shape()
  that relies on old inputs_outputs_specs api (and thus need ForwardMode
  => cannot be imported)
- decomon.metrics: removed as rely on previous api and will fail to
  import (need ForwardMode)
- decomon.wrappers: kept but untested. Need to be adapted
- decomon.wrappers_with_tuning: removed as need decomon.metrics
- decomon.models.crown: removed. Specific layers for conversion (like
  ReduceCrown, Fuse, ...) are now in dedicated modules within
  decomon.layers
