+++ src/decomon/layers/fuse.py
@@ -7,15 +7,15 @@ import keras
 from keras import ops as K
 from keras.layers import Layer
 
-from decomon.core import (
+from decomon.constants import Propagation
+from decomon.keras_utils import add_tensors, batch_multid_dot
+from decomon.layers.inputs_outputs_specs import InputsOutputsSpec
+from decomon.perturbation_domain import (
     BoxDomain,
-    InputsOutputsSpec,
     PerturbationDomain,
-    Propagation,
     get_lower_box,
     get_upper_box,
 )
-from decomon.keras_utils import add_tensors, batch_multid_dot
 from decomon.types import BackendTensor, Tensor
