--- src/decomon/metrics/__init__.py
@@ -1,10 +1,0 @@
-from .loss import (
-    build_asymptotic_crossentropy_model,
-    build_crossentropy_model,
-    build_radius_robust_model,
-)
-from .metric import (
-    build_formal_adv_check_model,
-    build_formal_adv_model,
-    build_formal_upper_model,
-)
