+++ tests/lirpa_comparison/test_comparison_lirpa.py
@@ -9,7 +9,7 @@ from numpy.testing import assert_almost_equal
 from onnx2keras import onnx_to_keras
 from onnx2torch import convert
 
-from decomon.core import ConvertMethod
+from decomon.constants import ConvertMethod
 from decomon.models.convert import clone
