+++ tests/test_convert_forward.py
@@ -2,7 +2,7 @@ import pytest
 from keras.models import Model
 from pytest_cases import fixture, parametrize
 
-from decomon.core import Propagation, Slope
+from decomon.constants import Propagation, Slope
 from decomon.layers.activations.activation import DecomonBaseActivation
 from decomon.models.forward_cloning import convert_forward
